+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2024-10-14 - Upgrade: Adopt changes from MDL-81031 and use new \core\user class.
 * 2024-10-07 - Prepare compatibility for Moodle 4.5.
 
 ### v4.4-r1